# src/main/java/dev/eperrone/atms/api/AtmsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package dev.eperrone.atms.api;

import dev.eperrone.atms.model.ATM;
import dev.eperrone.atms.model.SearchATMs200Response;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "ATMs", description = "Gestión de cajeros automáticos")
public interface AtmsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /atms/{atmId} : Obtener detalle de un cajero
     * Retorna los datos completos de un cajero automático por su ID.
     *
     * @param atmId ID del cajero automático (required)
     * @return Detalle del cajero obtenido exitosamente (status code 200)
     *         or No autorizado (status code 401)
     *         or Cajero no encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getATMById",
        summary = "Obtener detalle de un cajero",
        description = "Retorna los datos completos de un cajero automático por su ID.",
        tags = { "ATMs" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Detalle del cajero obtenido exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ATM.class))
            }),
            @ApiResponse(responseCode = "401", description = "No autorizado"),
            @ApiResponse(responseCode = "404", description = "Cajero no encontrado"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/atms/{atmId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ATM> getATMById(
        @Parameter(name = "atmId", description = "ID del cajero automático", required = true, in = ParameterIn.PATH) @PathVariable("atmId") Integer atmId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2025-01-15T14:30:00Z\", \"bank\" : \"Banco Nación\", \"address\" : \"Av. Mitre 123, Florida\", \"latitude\" : -34.5333, \"postalCode\" : \"B1605\", \"id\" : 12345, \"localityId\" : 5678, \"hasDeposit\" : false, \"longitude\" : -58.4833, \"has24h\" : true, \"status\" : \"active\", \"updatedAt\" : \"2025-03-08T22:45:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /atms/list : Listar cajeros paginados
     * Obtiene una lista paginada de todos los cajeros (sin filtros avanzados).
     *
     * @param offset Número de elementos a saltar antes de empezar a devolver resultados. (optional, default to 0)
     * @param limit Cantidad máxima de elementos a devolver (máximo 100). (optional, default to 10)
     * @param status Estado de los registros a devolver. (optional)
     * @return Lista de cajeros obtenida exitosamente (status code 200)
     *         or No autorizado (status code 401)
     *         or No encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "listATMs",
        summary = "Listar cajeros paginados",
        description = "Obtiene una lista paginada de todos los cajeros (sin filtros avanzados).",
        tags = { "ATMs" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de cajeros obtenida exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchATMs200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "No autorizado"),
            @ApiResponse(responseCode = "404", description = "No encontrado"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/atms/list",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SearchATMs200Response> listATMs(
        @Min(0) @Parameter(name = "offset", description = "Número de elementos a saltar antes de empezar a devolver resultados.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "offset", required = false, defaultValue = "0") Integer offset,
        @Min(1) @Max(100) @Parameter(name = "limit", description = "Cantidad máxima de elementos a devolver (máximo 100).", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "10") Integer limit,
        @Parameter(name = "status", description = "Estado de los registros a devolver.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) String status
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"pagination\" : { \"totalItems\" : 150, \"offset\" : 0, \"limit\" : 10 }, \"atms\" : [ { \"createdAt\" : \"2025-01-15T14:30:00Z\", \"bank\" : \"Banco Nación\", \"address\" : \"Av. Mitre 123, Florida\", \"latitude\" : -34.5333, \"postalCode\" : \"B1605\", \"id\" : 12345, \"localityId\" : 5678, \"hasDeposit\" : false, \"longitude\" : -58.4833, \"has24h\" : true, \"status\" : \"active\", \"updatedAt\" : \"2025-03-08T22:45:00Z\" }, { \"createdAt\" : \"2025-01-15T14:30:00Z\", \"bank\" : \"Banco Nación\", \"address\" : \"Av. Mitre 123, Florida\", \"latitude\" : -34.5333, \"postalCode\" : \"B1605\", \"id\" : 12345, \"localityId\" : 5678, \"hasDeposit\" : false, \"longitude\" : -58.4833, \"has24h\" : true, \"status\" : \"active\", \"updatedAt\" : \"2025-03-08T22:45:00Z\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /atms/search : Búsqueda avanzada de cajeros
     * Permite buscar cajeros por banco, localidad, código postal, disponibilidad 24h, etc. Compatible con paginación y scroll infinito.
     *
     * @param offset Número de elementos a saltar antes de empezar a devolver resultados. (optional, default to 0)
     * @param limit Cantidad máxima de elementos a devolver (máximo 100). (optional, default to 10)
     * @param bank Nombre o código del banco (ej. \&quot;Banco Nación\&quot;, \&quot;BBVA\&quot;, \&quot;Santander\&quot;) (optional)
     * @param localityId ID de la localidad (relacionado con la API de localidades) (optional)
     * @param postalCode Código postal aproximado (optional)
     * @param has24h Filtrar solo cajeros con atención 24 horas (optional)
     * @param q Búsqueda libre (dirección, banco, localidad o barrio) (optional)
     * @param status Estado del registro del cajero (optional)
     * @return Respuesta exitosa (status code 200)
     *         or Parámetros de búsqueda inválidos (status code 400)
     *         or No autorizado (status code 401)
     *         or No encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "searchATMs",
        summary = "Búsqueda avanzada de cajeros",
        description = "Permite buscar cajeros por banco, localidad, código postal, disponibilidad 24h, etc. Compatible con paginación y scroll infinito.",
        tags = { "ATMs" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Respuesta exitosa", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SearchATMs200Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Parámetros de búsqueda inválidos"),
            @ApiResponse(responseCode = "401", description = "No autorizado"),
            @ApiResponse(responseCode = "404", description = "No encontrado"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/atms/search",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SearchATMs200Response> searchATMs(
        @Min(0) @Parameter(name = "offset", description = "Número de elementos a saltar antes de empezar a devolver resultados.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "offset", required = false, defaultValue = "0") Integer offset,
        @Min(1) @Max(100) @Parameter(name = "limit", description = "Cantidad máxima de elementos a devolver (máximo 100).", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "10") Integer limit,
        @Parameter(name = "bank", description = "Nombre o código del banco (ej. \"Banco Nación\", \"BBVA\", \"Santander\")", in = ParameterIn.QUERY) @Valid @RequestParam(value = "bank", required = false) String bank,
        @Parameter(name = "localityId", description = "ID de la localidad (relacionado con la API de localidades)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "localityId", required = false) Integer localityId,
        @Parameter(name = "postalCode", description = "Código postal aproximado", in = ParameterIn.QUERY) @Valid @RequestParam(value = "postalCode", required = false) String postalCode,
        @Parameter(name = "has24h", description = "Filtrar solo cajeros con atención 24 horas", in = ParameterIn.QUERY) @Valid @RequestParam(value = "has24h", required = false) Boolean has24h,
        @Parameter(name = "q", description = "Búsqueda libre (dirección, banco, localidad o barrio)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "q", required = false) String q,
        @Parameter(name = "status", description = "Estado del registro del cajero", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) String status
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"pagination\" : { \"totalItems\" : 150, \"offset\" : 0, \"limit\" : 10 }, \"atms\" : [ { \"createdAt\" : \"2025-01-15T14:30:00Z\", \"bank\" : \"Banco Nación\", \"address\" : \"Av. Mitre 123, Florida\", \"latitude\" : -34.5333, \"postalCode\" : \"B1605\", \"id\" : 12345, \"localityId\" : 5678, \"hasDeposit\" : false, \"longitude\" : -58.4833, \"has24h\" : true, \"status\" : \"active\", \"updatedAt\" : \"2025-03-08T22:45:00Z\" }, { \"createdAt\" : \"2025-01-15T14:30:00Z\", \"bank\" : \"Banco Nación\", \"address\" : \"Av. Mitre 123, Florida\", \"latitude\" : -34.5333, \"postalCode\" : \"B1605\", \"id\" : 12345, \"localityId\" : 5678, \"hasDeposit\" : false, \"longitude\" : -58.4833, \"has24h\" : true, \"status\" : \"active\", \"updatedAt\" : \"2025-03-08T22:45:00Z\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    ResponseEntity<SearchATMs200Response> listATMs(Integer offset, Integer limit);
}